requirements.txt
web3==7.14.1
google-generativeai==0.8.6
python-dotenv==1.2.1
PyPDF2==3.0.1
pillow==12.1.0